+++ python/cudf/cudf/bindings/avro.pyx
@@ -5,10 +5,12 @@
 # cython: embedsignature = True
 # cython: language_level = 3
 
-from .cudf_cpp cimport *
+from cudf.bindings.cudf_cpp cimport *
 from cudf.bindings.cudf_cpp import *
 from cudf.bindings.avro cimport reader as avro_reader
 from cudf.bindings.avro cimport reader_options as avro_reader_options
+from cudf.bindings.utils cimport *
+from cudf.bindings.utils import *
 from libc.stdlib cimport free
 from libcpp.vector cimport vector
 from libcpp.memory cimport unique_ptr
@@ -60,35 +62,18 @@ cpdef cpp_read_avro(filepath_or_buffer, columns=None, skip_rows=None,
         )
 
     # Read data into columns
-    cdef cudf_table table
+    cdef cudf_table c_out_table
     if skip_rows is not None:
-        table = reader.get().read_rows(
+        c_out_table = reader.get().read_rows(
             skip_rows,
             num_rows if num_rows is not None else 0
         )
     elif num_rows is not None:
-        table = reader.get().read_rows(
+        c_out_table = reader.get().read_rows(
             skip_rows if skip_rows is not None else 0,
             num_rows
         )
     else:
-        table = reader.get().read_all()
+        c_out_table = reader.get().read_all()
 
-    # Extract read columns
-    outcols = []
-    new_names = []
-    cdef gdf_column* column
-    for i in range(table.num_columns()):
-        column = table.get_column(i)
-        data_mem, mask_mem = gdf_column_to_column_mem(column)
-        outcols.append(Column.from_mem_views(data_mem, mask_mem))
-        new_names.append(column.col_name.decode())
-        free(column.col_name)
-        free(column)
-
-    # Construct dataframe from columns
-    df = DataFrame()
-    for k, v in zip(new_names, outcols):
-        df[k] = v
-
-    return df
+    return table_to_dataframe(&c_out_table)