+++ python/cudf/cudf/bindings/stream_compaction.pyx
@@ -5,17 +5,13 @@
 # cython: embedsignature = True
 # cython: language_level = 3
 
-from libc.stdlib cimport free
-from libcpp.vector cimport vector
-
 from cudf.bindings.cudf_cpp cimport *
 from cudf.bindings.cudf_cpp import *
 from cudf.bindings.utils cimport *
+from cudf.bindings.utils import *
+
 from cudf.bindings.copying cimport cols_view_from_cols, free_table
 from cudf.bindings.copying import clone_columns_with_size
-from cudf.dataframe.column import Column
-from cudf.dataframe.buffer import Buffer
-
 from cudf.bindings.stream_compaction cimport *
 
 
@@ -59,21 +55,16 @@ def apply_drop_duplicates(in_index, in_cols, subset=None, keep='first'):
         key_cols = cols_view_from_cols(subset)
         key_table = new cudf_table(key_cols, len(subset))
 
-    cdef cudf_table out_table
+    cdef cudf_table c_out_table
     with nogil:
-        out_table = drop_duplicates(
-            c_in_table[0], key_table[0], keep_first
-        )
+        c_out_table = drop_duplicates(c_in_table[0], key_table[0], keep_first)
 
     free_table(key_table, key_cols)
     free_table(c_in_table, c_in_cols)
 
     # convert table to columns, index
-    out_cols = [
-        Column.from_mem_views(
-            *gdf_column_to_column_mem(i)
-        ) for i in out_table
-    ]
+    out_cols = columns_from_table(&c_out_table)
+
     return (out_cols[:-1], out_cols[-1])
 
 
@@ -97,8 +88,8 @@ def apply_apply_boolean_mask(cols, mask):
     with nogil:
         c_out_table = apply_boolean_mask(c_in_table[0], c_mask_col[0])
 
-    free(c_in_table)
-    free(c_mask_col)
+    free_table(c_in_table)
+    free_column(c_mask_col)
 
     return columns_from_table(&c_out_table)
 
@@ -145,8 +136,8 @@ def apply_drop_nulls(cols, how="any", subset=None, thresh=None):
         c_out_table = drop_nulls(c_in_table[0], c_keys_table[0],
                                  c_keep_threshold)
 
-    free(c_in_table)
-    free(c_keys_table)
+    free_table(c_in_table)
+    free_table(c_keys_table)
 
     result_cols = columns_from_table(&c_out_table)