+++ python/cudf/cudf/groupby/groupby.py
@@ -2,6 +2,7 @@
 
 import collections
 import itertools
+import pickle
 
 import cudf
 from cudf import MultiIndex
@@ -63,6 +64,21 @@ class SeriesGroupBy(_Groupby):
             obj=self._sr, by=by, level=level, sort=sort
         )
 
+    def serialize(self):
+        header = {}
+        header["groupby"] = pickle.dumps(self)
+        header["type"] = pickle.dumps(type(self))
+        header["sr"], frames = self._sr.serialize()
+        return header, frames
+
+    @classmethod
+    def deserialize(cls, header, frames):
+        groupby = pickle.loads(header["groupby"])
+        sr_typ = pickle.loads(header["sr"]["type"])
+        sr = sr_typ.deserialize(header["sr"], frames)
+        groupby._sr = sr
+        return groupby
+
     def _apply_aggregation(self, agg):
         return self._groupby.compute_result(agg)
 
@@ -118,6 +134,21 @@ class DataFrameGroupBy(_Groupby):
         raise AttributeError(
             "'DataFrameGroupBy' object has no attribute " "'{}'".format(key)
         )
+
+    def serialize(self):
+        header = {}
+        header["groupby"] = pickle.dumps(self)
+        header["type"] = pickle.dumps(type(self))
+        header["df"], frames = self._df.serialize()
+        return header, frames
+
+    @classmethod
+    def deserialize(cls, header, frames):
+        groupby = pickle.loads(header["groupby"])
+        df_typ = pickle.loads(header["df"]["type"])
+        df = df_typ.deserialize(header["df"], frames)
+        groupby._df = df
+        return groupby
 
 
 class _GroupbyHelper(object):
@@ -333,7 +364,8 @@ class _GroupbyHelper(object):
             return MultiIndex(
                 source_data=dataframe_from_columns(
                     key_columns, columns=key_names
-                )
+                ),
+                names=key_names,
             )
 
     def compute_result_column_index(self):