+++ python/dask_cudf/dask_cudf/io/tests/test_parquet.py
@@ -3,6 +3,7 @@ import os
 import pandas as pd
 import pytest
 
+import dask
 import dask.dataframe as dd
 from dask.dataframe.utils import assert_eq
 from dask.utils import natural_sort_key
@@ -15,7 +16,8 @@ df = pd.DataFrame(
     {
         "x": [i * 7 % 5 for i in range(nrows)],  # Not sorted
         "y": [i * 2.5 for i in range(nrows)],
-    }
+    },
+    index=pd.Index(range(nrows), name="index"),
 )  # Sorted
 ddf = dd.from_pandas(df, npartitions=npartitions)
 
@@ -60,12 +62,19 @@ def test_roundtrip_from_dask(tmpdir):
 
 def test_roundtrip_from_pandas(tmpdir):
     fn = str(tmpdir.join("test.parquet"))
+
+    # First without specifying an index
     dfp = df.copy()
-    dfp.index.name = "index"
-    dfp.to_parquet(fn, engine="pyarrow")
-    ddf2 = dask_cudf.read_parquet(fn, index="index")
-    # Losing the index name here for some reason
-    assert_eq(dfp, ddf2, check_index=False)
+    dfp.to_parquet(fn, engine="pyarrow", index=False)
+    dfp = dfp.reset_index(drop=True)
+    ddf2 = dask_cudf.read_parquet(fn)
+    assert_eq(dfp, ddf2, check_index=True)
+
+    # Now, specifying an index
+    dfp = df.copy()
+    dfp.to_parquet(fn, engine="pyarrow", index=True)
+    ddf2 = dask_cudf.read_parquet(fn, index=["index"])
+    assert_eq(dfp, ddf2, check_index=True)
 
 
 def test_strings(tmpdir):
@@ -79,6 +88,32 @@ def test_strings(tmpdir):
     ddf2.to_parquet(fn, engine="pyarrow")
     read_df = dask_cudf.read_parquet(fn, index=["a"])
     assert_eq(ddf2, read_df.compute().to_pandas())
+
+    read_df_cats = dask_cudf.read_parquet(
+        fn, index=["a"], strings_to_categorical=True
+    )
+    assert_eq(read_df_cats.dtypes, read_df_cats.compute().dtypes)
+    assert_eq(read_df_cats.dtypes[0], "int32")
+
+
+def test_dask_timeseries_from_pandas(tmpdir):
+
+    fn = str(tmpdir.join("test.parquet"))
+    ddf2 = dask.datasets.timeseries(freq="D")
+    pdf = ddf2.compute()
+    pdf.to_parquet(fn, engine="pyarrow")
+    read_df = dask_cudf.read_parquet(fn)
+    assert_eq(ddf2, read_df.compute().to_pandas())
+
+
+def test_dask_timeseries_from_dask(tmpdir):
+
+    fn = str(tmpdir)
+    ddf2 = dask.datasets.timeseries(freq="D")
+    ddf2.to_parquet(fn, engine="pyarrow")
+    read_df = dask_cudf.read_parquet(fn)
+    # Note: Loosing the index name here
+    assert_eq(ddf2, read_df.compute().to_pandas(), check_index=False)
 
 
 @pytest.mark.parametrize("index", [False, True])